+experiments Epoch
add experiment epoch 50-1500. The accuracy of the model will decrease if it has reached 550+ epochs. For the 600th epoch the accuracy results dropped to 87.61% and continued to drop significantly until the last known accuracy at the 1500th epoch was 38.67%. Although sometimes there is still an increase but the final result continues to fall.